% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logigram.R
\name{cora_dnf}
\alias{cora_dnf}
\title{Disjunctive normal form of a solution}
\usage{
cora_dnf(x, ...)
}
\arguments{
\item{x}{A solution from [cora_irredundant_sums()] or
[cora_irredundant_systems()].}

\item{...}{Unused.}
}
\value{
A character vector with one entry per outcome.
}
\description{
Renders a solution in the `"A*B+c<=>F"` notation that [cora_logigram()]
draws.
}
\examples{
df <- data.frame(A = c(1, 0, 1, 0), B = c(1, 0, 0, 1),
                 C = c(0, 1, 1, 0), OUT = c(1, 1, 0, 1))
cora_dnf(cora_irredundant_sums(cora_context(df, "OUT"))[[1]])
}
