% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cora_describe}
\alias{cora_describe}
\title{Descriptive rendering of a solution}
\usage{
cora_describe(x, cov = 1, ...)
}
\arguments{
\item{x}{A solution from [cora_irredundant_sums()] or
[cora_irredundant_systems()].}

\item{cov}{Minimum coverage score for the relation to be read as
necessary as well as sufficient.}

\item{...}{Unused.}
}
\value{
A character string.
}
\description{
Renders a solution as a sufficiency, necessity or equivalence statement,
depending on whether its inclusion and coverage scores clear the
thresholds of the analysis.
}
\examples{
df <- data.frame(A = c(1, 0, 1, 0), B = c(1, 0, 0, 1),
                 C = c(0, 1, 1, 0), OUT = c(1, 1, 0, 1))
sums <- cora_irredundant_sums(cora_context(df, "OUT"))
cora_describe(sums[[1]])
}
