% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systems.R
\name{cora_irredundant_systems}
\alias{cora_irredundant_systems}
\title{Irredundant systems of a multi-outcome analysis}
\usage{
cora_irredundant_systems(ctx)
}
\arguments{
\item{ctx}{A [cora_context()] with more than one outcome.}
}
\value{
A list of systems, of class `cora_systems`.
}
\description{
Solves the prime implicant chart of every outcome and combines the
single-outcome solutions into irredundant systems. Individual functions
inside a system need not be irredundant, but the system as a whole is.
}
\examples{
df <- data.frame(A = c(1, 1, 0, 0), B = c(2, 1, 2, 2), C = c(0, 1, 1, 2),
                 D = c(1, 0, 0, 0), OUT1 = c(1, 2, 0, 1),
                 OUT2 = c(2, 0, 1, 1), OUT3 = c(1, 0, 2, 1))
ctx <- cora_context(df, c("OUT1{1,2}", "OUT2{1}", "OUT3{1,0}"),
                    algorithm = "ON-OFF")
cora_irredundant_systems(ctx)
}
