% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_minaret}
\alias{swiss_minaret}
\title{Swiss minaret referendum}
\format{
A data frame with 11 rows and 6 columns: the conditions `A`, `L`,
  `S` and `T`, and the outcomes `X` and `M`.
}
\source{
Shipped with the Python CORA package,
  <https://github.com/PoliUniLu/cora>.
}
\usage{
swiss_minaret
}
\description{
Cantonal data on the 2009 Swiss referendum on the construction of
minarets, used as the multi-outcome example of the CORA package.
}
\examples{
ctx <- cora_context(swiss_minaret, c("X", "M"), algorithm = "ON-OFF")
cora_irredundant_systems(ctx)
}
\keyword{datasets}
